+++ packages/web/src/tabs/index.js
@@ -23,6 +23,7 @@ import * as ChangelogTab from './ChangelogTab.svelte';
 import * as DiagramTab from './DiagramTab.svelte';
 import * as DbKeyDetailTab from './DbKeyDetailTab.svelte';
 import * as QueryDataTab from './QueryDataTab.svelte';
+import * as ConnectionTab from './ConnectionTab.svelte';
 
 export default {
   TableDataTab,
@@ -50,4 +51,5 @@ export default {
   DiagramTab,
   DbKeyDetailTab,
   QueryDataTab,
+  ConnectionTab,
 };